requirements.txt
annotated-types==0.7.0
anyio==4.12.0
attrs==23.2.0
Babel==2.10.3
bcc==0.29.1
blinker==1.9.0
Brlapi==0.8.5
certifi==2023.11.17
chardet==5.2.0
charset-normalizer==3.4.4
click==8.1.6
cloud-init==25.1.4
colorama==0.4.6
command-not-found==0.3
configobj==5.0.8
cryptography==41.0.7
cupshelpers==1.0
dbus-python==1.3.2
defer==1.0.6
distro==1.9.0
distro-info==1.7+build1
docx==0.2.4
faiss-cpu==1.13.0
Flask==3.1.2
flask-cors==6.0.1
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.20.4
httpx==0.28.1
idna==3.6
itsdangerous==2.2.0
Jinja2==3.1.2
jiter==0.12.0
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.10.3
language-selector==0.1
launchpadlib==1.11.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
louis==3.29.0
lxml==6.0.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
netaddr==0.8.0
netifaces==0.11.0
numpy==2.3.5
oauthlib==3.2.2
olefile==0.46
openai==2.8.1
packaging==25.0
pandas==2.3.3
pdfminer.six==20251107
pdfplumber==0.11.8
pexpect==4.9.0
pillow==10.2.0
ptyprocess==0.7.0
pycairo==1.25.1
pycups==2.0.1
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.17.2
PyGObject==3.48.2
PyJWT==2.7.0
pyparsing==3.1.1
pypdfium2==5.1.0
pyrsistent==0.20.0
pyserial==3.5
pytesseract==0.3.13
python-apt==2.7.7+ubuntu3
python-dateutil==2.8.2
python-debian==0.1.49+ubuntu2
python-docx==1.2.0
python-dotenv==1.2.1
pytz==2024.1
pyxdg==0.28
PyYAML==6.0.1
requests==2.31.0
rich==13.7.1
setuptools==68.1.2
six==1.16.0
sniffio==1.3.1
systemd-python==235
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.2
unattended-upgrades==0.1
urllib3==2.0.7
wadllib==1.3.6
Werkzeug==3.1.4
wheel==0.42.0
xdg==5
xkit==0.0.0
gunicorn
flask-cors
faiss-cpu
pytesseract
pdfplumber
python-docx
python-dotenv
